+++ index.js
@@ -1,8 +1,6 @@
 var request = require('superagent')
 var serialize = require('form-serialize')
-var domify = require('domify')
 var Event = require('compose-event')
-var Dialog = require('compose-dialog')
 var slice = Array.prototype.slice
 
 var counter = 0
@@ -11,11 +9,11 @@ var callbacks = {}
 
 var Form = {
   listen: function(){
-    this.resetCallbacks()
+    self.resetCallbacks()
 
-    Event.on(document, 'submit', 'form[data-remote]',  this.submit.bind(this))
-    Event.on(document, 'submit', 'form', this.disableWith)
-    Event.on(document, 'click', 'a[data-method], a[data-confirm], button[data-method], button[data-confirm]', this.click.bind(this))
+    Event.on(document, 'submit', 'form[data-remote]',  self.submit)
+    Event.on(document, 'submit', 'form', self.disableWith)
+    Event.on(document, 'click', 'a[data-method], a[data-confirm], button[data-method], button[data-confirm]', self.click)
   },
 
   submit: function(event){
@@ -38,24 +36,47 @@ var Form = {
     else
       req.set('Accept', '*/*;q=0.5, text/javascript, application/javascript, application/ecmascript, application/x-ecmascript')
 
-    this.fireCallbacks(form, 'beforeSend', [req])
+    self.fireCallbacks(form, 'beforeSend', [req])
 
-    req.end(function(error, response) { this.handleResponse(form, error, response, currentRequest) }.bind(this))
+    req.end(function(error, response) { self.handleResponse(form, error, response, currentRequest) })
 
     return req
   },
 
   handleResponse: function(form, error, response, currentRequest){
-    if (error)
-      this.fireCallbacks(form, 'error', [currentRequest.xhr, currentRequest.xhr.status, error])
-    else if (response.error)
-      this.fireCallbacks(form, 'error', [currentRequest.xhr, response.status, response.error])
-    else
-      this.fireCallbacks(form, 'success', [response.body, response.status, currentRequest.xhr])
+    var xhr = currentRequest.xhr
+    var status = self.processStatus(xhr, response)
 
-    this.enableButtons(form)
-    this.fireCallbacks(form, 'complete', [currentRequest.xhr, response ? response.status : 0])
+    if (error || response.error) {
+      error = self.processErrorResponse(xhr, error || response.error)
+      self.fireCallbacks(form, 'error', [xhr, status, error])
+    } else {
+      self.fireCallbacks(form, 'success', [response.body, status, xhr])
+    }
+
+    self.enableButtons(form)
+    self.fireCallbacks(form, 'complete', [xhr, status])
     delete currentRequest
+  },
+
+  processStatus: function(xhr, response) {
+    if (response && response.status)
+      return response.status
+    else if (xhr && xhr.status) 
+      return xhr.status
+    else
+      return 0
+  },
+
+  processErrorResponse: function(xhr, message) {
+    if (xhr && xhr.response && xhr.repsonse.errors) {
+      return xhr.response.errors
+    }
+    try {
+      message = JSON.parse(xhr.response.text).messages
+    } finally {
+      return message
+    }
   },
 
   fireCallbacks: function(form, type, args) {
@@ -126,26 +147,36 @@ var Form = {
     })
   },
 
+  parentForm: function(el) {
+    while (el && el.tagName != "FORM") {
+      el = el.parentNode
+    }
+
+    return el
+  },
+
   click: function(event) {
     event.preventDefault()
     var el = event.currentTarget
-    var form = null
+    var form = parentForm(el)
 
-    if (el.dataset.method) {
-      form = this.buildForm(el)
-      document.body.appendChild(form)
+    // If the form submission trigger is not inside of a form, create one
+    if (!form) {
+      document.body.insertAdjacentHTML('beforeend', self.buildForm(el))
+      form = document.body.lastChild
     }
 
     if (el.dataset.confirm) {
-      this.confirm(el, form)
+      self.triggerConfirm(el, form)
     } else {
       form.submit()
     }
+
   },
 
-  confirm: function(el, form) {
+  triggerConfirm: function(el, form) {
     var opts = {
-      title: el.dataset.confirm,
+      title: el.dataset.confirm || null,
       message: el.dataset.message || null,
       continue: el.dataset.continue || DEFAULT_CONTINUE_BUTTON
     }
@@ -162,7 +193,16 @@ var Form = {
       opts.submit = el.dataset.submit
     }
 
-    Dialog.show(opts)
+    self.confirm(opts)
+  },
+
+  confirm: function(options) {
+    if(confirm(options.message || options.title)) {
+      if (options.follow)
+        window.location = options.follow
+      else if (options.submit)
+        document.querySelector(options.submit).submit()
+    }
   },
 
   buildForm: function(el) {
@@ -170,12 +210,13 @@ var Form = {
     var csrfToken = document.querySelector('meta[name=csrf-token]')
     var csrfParam = document.querySelector('meta[name=csrf-param]')
 
-    var form = domify('<form id="link-'+ (++counter) +'" class="hidden" action="'+el.href+'" method="post"></form>')
-
-    form.appendChild(domify('<input name="_method" value="' + method + '" type="hidden">'))
+    var form = '<form id="link-'+ (++counter) +'" class="hidden" action="'+el.dataset.url || el.dataset.action || el.href+'" method="post">'
+    form += '<input name="_method" value="' + method + '" type="hidden">'
 
     if (csrfToken && csrfParam)
-      form.appendChild(domify('<input name="' + csrfParam.getAttribute('content') + '" value="' + csrfToken.getAttribute('content') + '" type="hidden" />'))
+      form += '<input name="' + csrfParam.getAttribute('content') + '" value="' + csrfToken.getAttribute('content') + '" type="hidden">'
+
+    form += "</form>"
 
     return form
   }