+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.2.0
+- Added support for responses formatted in the [JSON api errors standard](http://jsonapi.org/examples/#error-objects).
+- Factored out domify dependency in favor of `insertAdjacentHTML`.
+- Factored out compose-dialog dependency in favor of native confirm, which can be overridden with a custom confirm function.
+- Fixed: Forms with a data-confirm as their default submission button will trigger the confirm dialog too.
+
 ## 2.1.0
 - Now form event callbacks pass the form element as the first argument.
 - You can set callback on the document to trigger on events for any form.